feat: add aster integration
we can now run a basic graphql server for billing service, with tests data & schemas
