+++ billing/Cargo.toml
@@ -6,7 +6,10 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+common = { path = "../common" }
 async-graphql = { version = "5.0.9", features = ["dynamic-schema"]}
 async-graphql-axum = "5.0.9"
 axum = "0.6.18"
 tokio = { version = "1.28.2", features = ["macros", "rt-multi-thread"]}
+async-trait = "0.1.51"
+log = "0.4.18"
