+++ billing/src/lib.rs
@@ -1,34 +1,2 @@
-mod models;
-
-use async_graphql::{http::GraphiQLSource, EmptyMutation, EmptySubscription, Schema};
-use async_graphql_axum::{GraphQLRequest, GraphQLResponse};
-use axum::{
-    extract::Extension,
-    response::{self, IntoResponse},
-    routing::get,
-    Router, Server,
-};
-use models::{BillingSchema, QueryRoot};
-
-async fn graphql_handler(schema: Extension<BillingSchema>, req: GraphQLRequest) -> GraphQLResponse {
-    schema.execute(req.into_inner()).await.into()
-}
-
-async fn graphiql() -> impl IntoResponse {
-    response::Html(GraphiQLSource::build().endpoint("/").finish())
-}
-
-pub async fn run() {
-    let schema = Schema::new(QueryRoot, EmptyMutation, EmptySubscription);
-
-    let app = Router::new()
-        .route("/", get(graphiql).post(graphql_handler))
-        .layer(Extension(schema));
-
-    println!("GraphiQL IDE: http://localhost:8000");
-
-    Server::bind(&"127.0.0.1:8000".parse().unwrap())
-        .serve(app.into_make_service())
-        .await
-        .unwrap();
-}
+mod schemas;
+pub mod services;
